Issue #229. Fix issue involving moderator.onSync sending app sync event to remote clients. This issue is that onSync would send the sync event before the sync event that triggered the onSync callback, causing the OP engine to crash.
Session setup:
Moderator
Client1
Client2

* Client1 sends sync to server.
* onMessage receives the sync, calls moderator.onSync
* Moderator sends sync event, and the server would publish this sync
  immediately.
* onSync returns control to onMessage which finally returns true,
  triggering the sending of the original Client1 sync to Client2.

The end result is Client2 sees the moderator sync before Client1 sync
and causes the OP engine to crash. This may in fact be an OP engine
issue (I was under the impression that the OP engine could always sort
out out-of-order message, since network latency may cause messages to be
received out of order, but I might be wrong.).